+++ app/locals.js
@@ -12,8 +12,16 @@ module.exports =
    * @param {NextFunction} next
    */
   (req, res, next) => {
-    res.locals.BASE_URL = config.baseURL
-    res.locals.ENVIRONMENT = config.env
+    let { path: basePath } = req
+
+    // Remove trailing slash
+    if (basePath.endsWith('/')) {
+      basePath = basePath.slice(0, -1)
+    }
+
+    res.locals.basePath = basePath
+    res.locals.baseUrl = config.baseUrl
+    res.locals.env = config.env
 
     // Adobe Analytics
     res.locals.adobeTrackingUrl = config.adobeTrackingUrl